updates edge class
