--- src/proximityDetection.html
@@ -1,9 +1,0 @@
-<body>
-
-<canvas id="canvas" width="800" height="600"></canvas>
-
-<button id="drawEdges">Draw Edges</button>
-
-<script src="edgeUtils.js"></script>
-
-</body>
